MAD/requirements.txt
cdlib
leidenalg
python-igraph
networkx
numpy
scipy
pandas
scikit-learn
tqdm
matplotlib
python-louvain
karateclub
